+++ controller/_autoload.php
@@ -5,3 +5,4 @@ namespace Technical_penguins\Newslurp\Controller;
 require_once('Database.php');
 require_once('Page.php');
 require_once('Story.php');
+require_once('User.php');