+++ model/_autoload.php
@@ -4,4 +4,3 @@ namespace Technical_penguins\Newslurp\Model;
 
 require_once('Email.php');
 require_once('Story.php');
-require_once('User.php');