+++ trigger.php
@@ -1,12 +1,11 @@
 <?php
 namespace Technical_penguins\Newslurp;
 
-session_start();
-
-require_once(__DIR__ . '/../_autoload.php');
+require_once(__DIR__ . '/_autoload.php');
 
 use Flight;
 use Technical_penguins\Newslurp\Action\Gmail;
+Use Technical_penguins\Newslurp\Controller\User;
 
 $gmail = new Gmail();
 Flight::redirect('/');