+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.1.0
+##MODIFIED
+- Removed GMail's dependency on session for values
+- Moved trigger outside the public, will not rely on HTML connection
+- User a controller, not a model
+- User now has a couple of functions for loading auth values for gmail
+
 # 1.0.3
 ## ADDED
 - Trigger.php (because Flight is not playing well with curl or wget)